Moved least connections and Tracing to load_balancing:selection
