+++ routini/src/load_balancing/selection/least_connections.rs
@@ -7,7 +7,7 @@ use std::{
 };
 
 use arc_swap::ArcSwap;
-use pingora::protocols::l4::socket::SocketAddr;
+use pingora::{protocols::l4::socket::SocketAddr, upstreams::peer::Tracing};
 use smallvec::SmallVec;
 
 use crate::load_balancing::{
@@ -104,6 +104,31 @@ impl BackendIter for LeastConnectionsIter {
         } else {
             None
         }
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct ConnectionsTracer(pub SocketAddr);
+
+impl Tracing for ConnectionsTracer {
+    fn on_connected(&self) {
+        CONNECTIONS
+            .load()
+            .get(&self.0)
+            .and_then(|(_, count)| Some(count.fetch_add(1, Ordering::Relaxed)));
+        log::debug!("incremented connection {}", &self.0)
+    }
+
+    fn on_disconnected(&self) {
+        CONNECTIONS
+            .load()
+            .get(&self.0)
+            .and_then(|(_, count)| Some(count.fetch_sub(1, Ordering::Relaxed)));
+        log::debug!("decremented connection {}", &self.0);
+    }
+
+    fn boxed_clone(&self) -> Box<dyn Tracing> {
+        Box::new(self.clone()) as Box<dyn Tracing>
     }
 }
